requirements.txt
certifi==2018.1.18
Django==2.0.4
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
PyJWT==1.6.1
PyMySQL==0.8.0
pytz==2018.4
wincertstore==0.2
